requirements.txt
Django == 4.0
djangorestframework
markdown
django-filter

